[NEW]: meta
The meta parser is now available. A few modifications had to be done to
several node parsers. In particular we should pay attention to
`element.py`. When adding a property to a node we should append `__` to
the property so that the property name will not get confused with the
element attribute. It should be noted that the property `pos` appears
in many styles. These will be trouble if a user decides to use `pos` as
an element attribute.