[UPDATE]: meta
Meta keywords are case-insensitive and may consist of letters, numbers,
underscores and dashes and must end with a colon. This update will
break several other things that depend on this module since each entry
stores items instead of just plain text. This allows us to provide a
list of values to each keyword. This update also fixes failing tests
with the reference module.